+++ src/components/SelectLang/index.tsx
@@ -1,0 +1,51 @@
+import React from 'react';
+import { formatMessage, setLocale, getLocale } from 'umi-plugin-locale';
+import { Menu, Icon } from 'antd';
+import classNames from 'classnames';
+import HeaderDropdown from '../HeaderDropdown';
+import styles from './index.less';
+
+interface ISelectLangProps {
+  className: string;
+}
+const SelectLang: React.SFC<ISelectLangProps> = props => {
+  const { className } = props;
+  const selectedLang = getLocale();
+  const changeLang = ({ key }) => {
+    setLocale(key);
+  };
+  const locales = ['zh-CN', 'zh-TW', 'en-US', 'pt-BR'];
+  const languageLabels = {
+    'zh-CN': '简体中文',
+    'zh-TW': '繁体中文',
+    'en-US': 'English',
+    'pt-BR': 'Português',
+  };
+  const languageIcons = {
+    'zh-CN': '🇨🇳',
+    'zh-TW': '🇭🇰',
+    'en-US': '🇬🇧',
+    'pt-BR': '🇧🇷',
+  };
+  const langMenu = (
+    <Menu className={styles.menu} selectedKeys={[selectedLang]} onClick={changeLang}>
+      {locales.map(locale => (
+        <Menu.Item key={locale}>
+          <span role="img" aria-label={languageLabels[locale]}>
+            {languageIcons[locale]}
+          </span>{' '}
+          {languageLabels[locale]}
+        </Menu.Item>
+      ))}
+    </Menu>
+  );
+  return (
+    <HeaderDropdown overlay={langMenu} placement="bottomRight">
+      <span className={classNames(styles.dropDown, className)}>
+        <Icon type="global" title={formatMessage({ id: 'navBar.lang' })} />
+      </span>
+    </HeaderDropdown>
+  );
+};
+
+export default SelectLang;
